Restructure into package with create_app
