+++ microstructure_server/routes/download.py
@@ -1,0 +1,14 @@
+from flask import Blueprint, send_file
+import os
+
+bp = Blueprint('download', __name__)
+
+@bp.route('/download_processed_data')
+def download_processed_data():
+    try:
+        file_path = "tmp/enhanced_ebsd_map.png"
+        if not os.path.exists(file_path):
+            return "Processed file not found.", 404
+        return send_file(file_path, mimetype='image/png', as_attachment=True, download_name="enhanced_ebsd_map.png")
+    except Exception:
+        return "Internal Server Error", 500
