+++ microstructure_server/routes/main.py
@@ -1,0 +1,11 @@
+from flask import Blueprint, render_template
+
+bp = Blueprint('main', __name__)
+
+@bp.route('/')
+def home():
+    return render_template('home.html')
+
+@bp.route('/help_faq')
+def help_faq():
+    return render_template('help_faq.html')
