+++ templates/base.html
@@ -12,7 +12,7 @@
 <body>
     <nav class="navbar navbar-expand-lg navbar-dark">
         <div class="container">
-            <a class="navbar-brand d-flex align-items-center" href="{{ url_for('home') }}">
+            <a class="navbar-brand d-flex align-items-center" href="{{ url_for('main.home') }}">
                 <span class="me-2"><i class="bi bi-atom"></i></span>
                 Microstructural Analysis
             </a>
@@ -22,12 +22,12 @@
             <div class="collapse navbar-collapse" id="navbarNav">
                 <ul class="navbar-nav ms-auto">
                     <li class="nav-item">
-                        <a class="nav-link" href="{{ url_for('home') }}">
+                        <a class="nav-link" href="{{ url_for('main.home') }}">
                             <i class="bi bi-house-door me-1"></i>Home
                         </a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link" href="{{ url_for('help_faq') }}">
+                        <a class="nav-link" href="{{ url_for('main.help_faq') }}">
                             <i class="bi bi-question-circle me-1"></i>Help & FAQ
                         </a>
                     </li>
@@ -42,7 +42,7 @@
                 <div>
                     <nav aria-label="breadcrumb">
                         <ol class="breadcrumb mb-0">
-                            <li class="breadcrumb-item"><a href="{{ url_for('home') }}">Home</a></li>
+                            <li class="breadcrumb-item"><a href="{{ url_for('main.home') }}">Home</a></li>
                             <li class="breadcrumb-item active" aria-current="page">{% block breadcrumb_title %}{{ self.title() }}{% endblock %}</li>
                         </ol>
                     </nav>
