+++ templates/home.html
@@ -16,7 +16,7 @@
                 <div class="card-body text-center">
                     <h5 class="card-title">Super Resolution</h5>
                     <p class="card-text">Enhance the resolution of your microstructural images using advanced AI algorithms.</p>
-                    <a href="{{ url_for('super_resolution') }}" class="btn btn-primary">Use Tool</a>
+                    <a href="{{ url_for('super_resolution.super_resolution') }}" class="btn btn-primary">Use Tool</a>
                 </div>
             </div>
         </div>
@@ -29,7 +29,7 @@
                 <div class="card-body text-center">
                     <h5 class="card-title">EBSD Clean-Up</h5>
                     <p class="card-text">Clean and process your EBSD data for better analysis and visualization.</p>
-                    <a href="{{ url_for('ebsd_cleanup') }}" class="btn btn-primary">Use Tool</a>
+                    <a href="{{ url_for('ebsd_cleanup.ebsd_cleanup') }}" class="btn btn-primary">Use Tool</a>
                 </div>
             </div>
         </div>
