+++ tests/conftest.py
@@ -6,10 +6,11 @@ import pytest
 # Add project root to sys.path
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 
-from app import app as flask_app  # Now it should work!
+from microstructure_server import create_app
 
 @pytest.fixture
 def client():
-    flask_app.config['TESTING'] = True
-    with flask_app.test_client() as client:
+    app = create_app()
+    app.config['TESTING'] = True
+    with app.test_client() as client:
         yield client
